FINALLY WORKING LOBBY
It took awhile. But i figured out the best unique way to start the server. Obviously if we're got any network scientist/engineers in the class they can VERY easy catch all the data flow cause i believe it isn't the most pythonic way of doing it, But it sorta works so YAY